+++ ft_strcat.c
@@ -6,7 +6,7 @@
 /*   By: jczech <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/12 13:47:03 by jczech            #+#    #+#             */
-/*   Updated: 2019/09/12 20:54:37 by jczech           ###   ########.fr       */
+/*   Updated: 2019/09/24 19:13:39 by jczech           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strcat(char *s1, const char *s2)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	j = 0;
 	i = ft_strlen(s1);
