+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: jczech <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/14 19:11:14 by jczech            #+#    #+#             */
-/*   Updated: 2019/09/14 19:47:53 by jczech           ###   ########.fr       */
+/*   Updated: 2019/09/24 19:31:41 by jczech           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_strrchr(const char *str, int ch)
 {
-	int	i;
+	int		i;
 
-	i = ft_strlen((char *)str);
+	i = (int)ft_strlen((char *)str);
 	while (str[i] != ch && i != -1)
 		i--;
 	if (str[i] == ch)
